[pf-579] Prevent sharing of a campaign
If the user still has questions to answer
Remove duplicate test file
Refactor various files
